+++ tests/tests/test_greedy.py
@@ -94,7 +94,7 @@ def test_greedy_secondary_clustering_1(self):
 
 def test_multiround_primary_clustering_2(self):
     """
-    Make sure this works with dereplicate
+    Make sure this works with dereplicate --ignoreGenomeQuality
     """
     test_dir = self.test_dir
 
@@ -117,6 +117,75 @@ def test_multiround_primary_clustering_2(self):
 
     # Make sure you still got the correct clustering
     Cdb = wd.get_db('Cdb').sort_values('genome').reset_index(drop=True)
+    assert test_utils.compare_dfs2(CdbS, Cdb, verbose=True)
+
+    # Make sure it handles plotting gracefully
+    drep.d_analyze.mash_dendrogram_from_wd(wd, plot_dir=test_dir)
+
+@pytest.mark.skip(reason="This test doesn't work; you don't know how to have clusterAlg impact the resulting Mdb")
+def test_multiround_primary_clustering_3(self):
+    """
+    Make sure multiround_primary_clustering respects "clusterAlg"
+    """
+    test_dir = self.test_dir
+
+    # Run it under normal conditions
+    args = drep.argumentParser.parse_args(['dereplicate', self.wd_loc, '--clusterAlg', 'single', '--primary_chunksize', '3', '--multiround_primary_clustering', '--ignoreGenomeQuality', '-pa', '0.75', '-d', '-g'] + self.genomes)
+    drep.controller.Controller().parseArguments(args)
+
+    # Load test results
+    wd = drep.WorkDirectory.WorkDirectory(self.wd_loc)
+    Mdb = wd.get_db('Mdb')
+    Cdb = wd.get_db('Cdb').sort_values('genome').reset_index(drop=True)
+    del Cdb['cluster_method']
+
+    # Run it with a different clusterAlg
+    shutil.rmtree(self.working_wd_loc)
+    args = drep.argumentParser.parse_args(['dereplicate', self.working_wd_loc, '--clusterAlg', 'complete', '--primary_chunksize', '3', '--multiround_primary_clustering', '--ignoreGenomeQuality', '-pa', '0.75', '-d', '-g'] + self.genomes)
+    drep.controller.Controller().parseArguments(args)
+
+    # Load test results
+    wd = drep.WorkDirectory.WorkDirectory(self.working_wd_loc)
+    Mdb2 = wd.get_db('Mdb')
+    Cdb2 = wd.get_db('Cdb').sort_values('genome').reset_index(drop=True)
+    del Cdb2['cluster_method']
+
+    # Make sure you get different clustering
+    assert not test_utils.compare_dfs2(Cdb, Cdb2, verbose=True)
+
+def test_multiround_primary_clustering_4(self):
+    """
+    Make sure this works with dereplicate without --ignoreGenomeQuality and with --SkipSecondary
+    """
+    test_dir = self.test_dir
+
+    # Run it under normal conditions
+    args = drep.argumentParser.parse_args(['dereplicate', self.wd_loc, '--primary_chunksize', '3', '--multiround_primary_clustering', '--ignoreGenomeQuality', '--SkipSecondary', '-d', '-g'] + self.genomes)
+    drep.controller.Controller().parseArguments(args)
+
+    # Load test results
+    wd = drep.WorkDirectory.WorkDirectory(self.wd_loc)
+
+    # Load solutions
+    wdS = drep.WorkDirectory.WorkDirectory(self.s_wd_loc)
+    CdbS = wdS.get_db('Cdb').sort_values('genome').reset_index(drop=True)
+
+    # Make sure you didn't pairwise
+    Mdb = wd.get_db('Mdb')
+    assert len(Mdb) != 25
+    assert 'genome_chunk' in list(Mdb.columns)
+    assert len(Mdb['genome_chunk'].unique()) == 3
+
+    # Make sure you still got the correct clustering
+    Cdb = wd.get_db('Cdb').sort_values('genome').reset_index(drop=True)
+
+    # Keep only the relevant columns
+    del CdbS['secondary_cluster']
+    del CdbS['comparison_algorithm']
+    del CdbS['threshold']
+    Cdb = Cdb[list(CdbS.columns)]
+    #print(CdbS.columns)
+
     assert test_utils.compare_dfs2(CdbS, Cdb, verbose=True)
 
     # Make sure it handles plotting gracefully